Moved GetLevelBunburrowStylePatch to Bunject Fixed namespaces of Yard patches
