+++ Bunject/Patches/LevelIndicatorGeneratorPatches.cs
@@ -1,4 +1,6 @@
-﻿using HarmonyLib;
+﻿using Bunburrows;
+using Bunject.Internal;
+using HarmonyLib;
 using Levels;
 using System;
 using System.Collections.Generic;
@@ -27,6 +29,18 @@ namespace Bunject.Patches.LevelIndicatorGeneratorPatches
           yield return instruction;
         }
       }
+    }
+  }
+  [HarmonyPatch(typeof(LevelIndicatorGenerator), nameof(LevelIndicatorGenerator.GetLevelBunburrowStyle))]
+  internal class GetLevelBunburrowStylePatch
+  {
+    public static BunburrowStyle Postfix(BunburrowStyle __result, LevelIdentity levelIdentity)
+    {
+      if (levelIdentity.Bunburrow.IsCustomBunburrow())
+      {
+        return AssetsManager.LevelsLists[levelIdentity.Bunburrow.ToBunburrowName()][levelIdentity.Depth].BunburrowStyle;
+      }
+      return __result;
     }
   }
 }
